Use shared crossbar helpers in crossbar testing
