+++ tests/integration/test_crossbar_integration.py
@@ -1,14 +1,12 @@
 import os
 import time
 import pytest
-import urllib.request
-
-from urllib.error import URLError
 
 import docker
-import numpy as np
 
 from ocs.matched_client import MatchedClient
+
+from util2 import create_crossbar_fixture, restart_crossbar
 
 try:
     from so3g import hk
@@ -16,6 +14,9 @@ except ModuleNotFoundError as e:
     print(f"Unable to import so3g: {e}")
 
 pytest_plugins = ("docker_compose",)
+
+
+wait_for_crossbar = create_crossbar_fixture()
 
 
 @pytest.mark.spt3g
@@ -30,43 +31,6 @@ def test_so3g_spt3g_import():
     # Just to prevent flake8 from complaining
     print(so3g.__file__)
 
-# Fixture to wait for crossbar server to be available.
-@pytest.fixture(scope="function")
-def wait_for_crossbar(function_scoped_container_getter):
-    """Wait for the crossbar server from docker-compose to become responsive."""
-    attempts = 0 
-
-    while attempts < 6:
-        try:
-            code = urllib.request.urlopen("http://localhost:8001/info").getcode()
-        except (URLError, ConnectionResetError):
-            print("Crossbar server not online yet, waiting 5 seconds.")
-            time.sleep(5)
-
-        attempts += 1
-
-    assert code == 200
-    print("Crossbar server online.")
-
-def restart_crossbar():
-    """Restart the crossbar server and wait for it to come back online."""
-    client = docker.from_env()
-    crossbar_container = client.containers.get('crossbar')
-    crossbar_container.restart()
-
-    attempts = 0
-
-    while attempts < 6:
-        try:
-            code = urllib.request.urlopen("http://localhost:8001/info").getcode()
-        except (URLError, ConnectionResetError):
-            print("Crossbar server not online yet, waiting 5 seconds.")
-            time.sleep(5)
-
-        attempts += 1
-
-    assert code == 200
-    print("Crossbar server online.")
 
 @pytest.mark.integtest
 def test_testing(wait_for_crossbar):
